api/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.15
langchain-core==0.2.37
langchain-mistralai==0.1.12
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.2
langsmith==0.1.109
multidict==6.0.5
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pinecone-client==3.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.7
